+++ common/Cypher.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "PipeMsg.h"
+#include <cstring>
 extern"C"
 {
 #include "utils/rc4.h"
@@ -11,26 +12,26 @@ class Cypher
 {
 public:
 
-	Cypher(EncryMethod  method, const char* key, uint32_t keyLen);
+	Cypher(EncryMethod  method, const char* key, size_t keyLen);
 
 	virtual ~Cypher();
 
-	char* encode(char* data, uint32_t len, uint32_t& outLen);
+	char* encode(char* data, size_t len, size_t& outLen);
 
-	char* decode(EncryMethod  method, char* data, uint32_t len, uint32_t& outLen);
+	char* decode(EncryMethod  method, char* data, size_t len, size_t& outLen);
 
 	EncryMethod getMethod();
 
 protected:
 
-	void resizeBuf(uint32_t len);
+	void resizeBuf(size_t len);
 
 private:
 	char* m_key;
-	uint32_t  m_keyLen;
+	size_t  m_keyLen;
 	EncryMethod m_method;
 
 	char* m_cacheBuf;
-	uint32_t m_cacheLen;
+	size_t m_cacheLen;
 	struct rc4_state m_state;
 };
